only 1 node_modules, refactored and renamed folders
Made it so there is only 1 node_modules folder, which should be in the root folder and not inside the frontend or backend. Refactored/Renamed React to frontend and server to backend. In the future when running any npm installs make sure you're in the PS C:\Users\jaime\Documents\GitHub\Redacted_Speed_Game> to keep only 1 node_modules folder.