ReadyUp for Game in Waiting Room
Added functionality for users to click ready and for other users to see they are ready and for them to see that they are waiting for other users to ready up.
Once all users are ready the countdown timer begins.
Also fixed a bug, where I forgot to socket.join users to rooms.
When timer ends users are directed to the game page.